extra credit
fancy terrain gen

+++ src/Water.java
@@ -1,7 +1,7 @@
 public class Water extends Terrain {
     public Water(){
         hue = (float) 0.7;
-        sat = 1;
+        sat = (float) 1;
         elevation = (float)Math.random();
     }
 }
